+++ lib/aspect_frontend_server.dart
@@ -1,0 +1,3 @@
+library aspect_frontend_server;
+
+export 'transformers.dart';