+++ test/aspect_frontend_server_test.dart
@@ -1,1 +1,4 @@
+import 'package:dydart/aspect_frontend_server.dart';
+import 'package:test/test.dart';
+
 void main() {}