+++ CHANGELOG.md
@@ -13,3 +13,31 @@
 ## 0.0.4
 
 * release for flutter 2.10.1
+
+## flutter2.10.1
+
+* update for flutter 2.10.1
+
+## flutter3.0.0
+
+* update for flutter 3.0.0
+* add try catch inject
+* add extension support
+
+
+## flutter3.1.0
+
+* update for flutter 3.1.0
+
+
+## flutter3.1.0.fix2
+
+* fix bug mixin not work
+* fix bug dispatchEvent inject not work
+* add dump_kernel.dart.snapshot
+* add mixin demo
+
+
+## flutter3.1.0.fix3
+* change folder 
+* old compile need update some code in dart-sdk,now change it instead